garbage_collector/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
charset-normalizer==3.1.0
Django==4.1.7
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fake-headers==1.0.2
frozenlist==1.3.3
-e git+https://github.com/liskunovich/garbage_collector.git@094aafd895d9304fec6592b9a220a6afa5cd1adb#egg=garbage_collector
html5lib==1.1
idna==3.4
Markdown==3.4.1
multidict==6.0.4
PyJWT==2.6.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.3
webencodings==0.5.1
yarl==1.8.2
